🐛 Fix dbus exception after gnome-shell recovers
Motivation:
Sometimes gnome-shell crashes but recovers immediately. In these
cases, the global singleton SCREENSAVER instance would be invalidated,
and this script would need to be reloaded in weechat to fix it.

Modification:
This change gets the ScreenSaver object from the dbus each time the
function is called, so it will always get the valid one.

Result:
This script now tolerates gnome-shell crashes.